实现了字符帧同步通信，这样单片机与单片机之间高速连续多帧通信成为可行的方案 比如以后可以实例化几个接收缓冲元件，几个发送元件，接受和发送元件数量可不同 甚至用数组的方式实现等等，这样就可以做到多指令同时输出执行。 单片机之间执行执行也可以连续高速执行。

+++ keilC51/modbus_rtu.c
@@ -14,11 +14,12 @@
 #include "sys_info.h"
 #include "modbus_rtu.h"
 #include "modbus_ascii.h"
+#include "compiler.h"
 
 
 unsigned int CRC16(unsigned char *Array,unsigned int Len)
 {
-	unsigned int  IX,IY,CRC;
+	WORD  IX,IY,CRC;
 	CRC=0xFFFF;//set all 1
 	if (Len<=0) {
 		CRC = 0;
@@ -26,7 +27,7 @@ unsigned int CRC16(unsigned char *Array,unsigned int Len)
 		Len--;
 		for (IX=0;IX<=Len;IX++)
 		{
-			CRC=CRC^(unsigned int)(Array[IX]);
+			CRC=CRC^(WORD)(Array[IX]);
 			for(IY=0;IY<=7;IY++) {
 				if ((CRC&1)!=0) {
 					CRC=(CRC>>1)^0xA001;
